[#71 & #68] refactor: move back errors to utils to differentiate with default type and package
