+++ pkg/utils/errors.go
@@ -1,4 +1,4 @@
-package error
+package utils
 
 import "errors"
